bg army 20200625

+++ django_relation/mysite/articles/views.py
@@ -2,13 +2,10 @@ from django.shortcuts import render, redirect, get_object_or_404
 # DVDH
 # django가 주는 views에서 쓸 decorators http를 위한
 from django.views.decorators.http import require_POST
-
 from django.contrib.auth.decorators import login_required
-
+from django.contrib import messages
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
-from IPython import embed
-
 
 # Create your views here.
 def index(request):
@@ -34,13 +31,15 @@ def detail(request, article_pk):
 @login_required
 def create(request):
     if request.method == "POST":
-        form = ArticleForm(request.POST)
-        
+        form = ArticleForm(request.POST, request.FILES)
         if form.is_valid():
             article = form.save(commit=False)
             article.user = request.user
             article.save()
-            return redirect('articles:detail',article.pk)
+            messages.success(request, '게시글 작성 완료!!!!!')
+            return redirect('articles:detail', article.pk)
+        else:
+            messages.error(request, '너 잘못된 데이터를 넣었어!!!')
     else:
         form = ArticleForm()
     context = {
@@ -53,7 +52,7 @@ def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     if article.user == request.user:
         if request.method == "POST":
-            form = ArticleForm(request.POST, instance=article)
+            form = ArticleForm(request.POST, request.FILES, instance=article)
             if form.is_valid():
                 article = form.save()
                 return redirect('articles:detail', article.pk)
@@ -64,39 +63,71 @@ def update(request, article_pk):
         }
         return render(request, 'articles/form.html', context)
     else:
-        return redirect('articles:index')
+        return redirect('articles:detail', article.pk)
 
 @require_POST
 def delete(request, article_pk):
     if request.user.is_authenticated:
         article = get_object_or_404(Article, pk=article_pk)
-        article.delete()
-    return redirect('articles:index')
+        if article.user == request.user:
+            article.delete()
+            return redirect('articles:index')
+        else:
+            return redirect('articles:detail', article.pk)
+    return redirect('articles:login')
 
 @require_POST
 def comment_create(request, article_pk):
-    # article = Article.objects.get(pk=article_pk)
-    article = get_object_or_404(Article, pk=article_pk)
-    comment_form = CommentForm(request.POST)
-    if comment_form.is_valid():
-        comment = comment_form.save(commit=False)
-        comment.user = request.user
-        comment.article_id = article.pk
-        comment.save()
-        return redirect('articles:detail', article.pk)
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=article_pk)
+        comment_form = CommentForm(request.POST)
+        if comment_form.is_valid():
+            comment = comment_form.save(commit=False)
+            comment.article = article
+            comment.user = request.user
+            comment.save()
+            return redirect('articles:detail', article.pk)
+        else:
+            context = {
+                'comment_form': comment_form,
+                'article': article
+            }
+            return render(request, 'articles/index.html', context)
     else:
-        context = {
-            'comment_form': comment_form,
-            'article': article
-        }
-    return redirect('articles:detail', context)
+        return redirect('accounts:login')
+    
 
 @require_POST
 def comment_delete(request, article_pk, comment_pk):
-    comment = get_object_or_404(Comment, pk=comment_pk)
-    if comment.user == request.user :
-        comment.delete()
+    if request.user.is_authenticated:
+        comment = get_object_or_404(Comment, pk=comment_pk)
+        if comment.user == request.user:
+            comment.delete()
         return redirect('articles:detail', article_pk)
-    return redirect('articles:detail', article_pk)
-    
-    
+    else:
+        return redirect('accounts:login')
+
+@login_required
+def like(request,article_pk):
+    #특정 게시글에 대한 정보
+    article = get_object_or_404(Article,pk=article_pk)
+    #좋아요를 누른 유저에 대한 정보
+    user = request.user
+    #사용자가 게시글의 좋아요 목록에 있으면 지우기
+    if user in article.like_users.all():
+        article.like_users.remove(user)
+    else :
+        article.like_users.add(user)
+
+    return redirect('articles:index')
+
+@login_required
+def recommend(request,article_pk):
+    article = get_object_or_404(Article,pk=article_pk)
+    user = request.user
+
+    if user in article.recommend_users.all():
+        article.recommend_users.remove(user)
+    else :
+        article.recommend_users.add(user)
+    return redirect('articles:index')
